feat(143): add student courses summary query
